refactor: directory clean up
