+++ src/main/java/com/com2here/com2hereback/dto/oauthinfo/GoogleInfo.java
@@ -1,5 +1,6 @@
-package com.com2here.com2hereback.dto;
+package com.com2here.com2hereback.dto.oauthinfo;
 
+import com.com2here.com2hereback.dto.oauthaccount.GoogleAccount;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
@@ -7,14 +8,14 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
-@ToString
 @Getter
 @NoArgsConstructor
+@ToString
 @JsonNaming(value = PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class KakaoInfo {
-    private KakaoAccount kakaoAccount;
+public class GoogleInfo {
+    private GoogleAccount googleAccount;
 
-    public static KakaoInfo fail() {
+    public static GoogleInfo fail() {
         return null;
     }
 }
