+++ src/main/java/com/com2here/com2hereback/dto/oauthinfo/KakaoInfo.java
@@ -1,5 +1,6 @@
-package com.com2here.com2hereback.dto;
+package com.com2here.com2hereback.dto.oauthinfo;
 
+import com.com2here.com2hereback.dto.oauthaccount.KakaoAccount;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
@@ -7,14 +8,14 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
+@ToString
 @Getter
 @NoArgsConstructor
-@ToString
 @JsonNaming(value = PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class NaverInfo {
-    private NaverAccount naverAccount;
+public class KakaoInfo {
+    private KakaoAccount kakaoAccount;
 
-    public static NaverInfo fail() {
+    public static KakaoInfo fail() {
         return null;
     }
 }
